update fitur hapus film per film , dan edit tapi masih di kembangkan

+++ src/components/atoms/FilmContentFourth.jsx
@@ -15,13 +15,23 @@ const FilmContentFourth = forwardRef((props, ref) => {
   }, []);
   const handleFilm = () => {
     let saveFilm = JSON.parse(localStorage.getItem("selectedFilm")) || [];
-    saveFilm.push(selectedFilm);
-    localStorage.setItem("selectedFilm", JSON.stringify(saveFilm));
-    Swal.fire({
-      title: "Success!!",
-      text: "Film berhasil ditambahkan 😁",
-      icon: "success",
-    });
+    const isFilmExist = saveFilm.some((film) => film.id === selectedFilm.id);
+
+    if (!isFilmExist) {
+      saveFilm.push(selectedFilm);
+      localStorage.setItem("selectedFilm", JSON.stringify(saveFilm));
+      Swal.fire({
+        title: "Success!!",
+        text: "Film berhasil ditambahkan!",
+        icon: "success",
+      });
+    } else {
+      Swal.fire({
+        title: "Oops!!",
+        text: "Film sudah ada didaftar!",
+        icon: "error",
+      });
+    }
   };
   return (
     <>
@@ -61,7 +71,7 @@ const FilmContentFourth = forwardRef((props, ref) => {
               <div className="flex flex-col ml-2">
                 <div className=" w-full flex justify-center">
                   <h3 className="text-lg font-bold text-gray-900 mb-2">
-                    {selectedFilm.judul}
+                    Judul : {selectedFilm.judul}
                   </h3>
                 </div>
                 <div className="">
